Create topological.py
Explanation (BFS Approach - Kahn’s Algorithm):
in-degree: This array keeps track of how many incoming edges each vertex has. Vertices with in-degree = 0 can be processed first.
Queue: Vertices with in-degree 0 are added to a queue and processed. As vertices are processed, the in-degrees of their neighbors are decreased. If a neighbor's in-degree becomes 0, it is added to the queue.
If the graph contains a cycle, topological sorting is not possible. @mevipinmaurya